% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Merge_PVR_Traits.R
\name{Merge_PVR_Traits}
\alias{Merge_PVR_Traits}
\title{Merge_PVR_Traits}
\usage{
Merge_PVR_Traits(my_tree, my_df, ID)
}
\arguments{
\item{my_tree}{phylogenetic tree}

\item{my_df}{traits data frame}

\item{ID}{Column name containing species names}
}
\description{
Merge eigenvectors to traits dataframe
}
