% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PVR_Robust.R
\name{PVR_Robust}
\alias{PVR_Robust}
\title{PVR_Robust}
\usage{
PVR_Robust(my_df, my_tree, ID, Y, X, method = "mMorI", estimator = "L2")
}
\arguments{
\item{my_df}{traits data frame}

\item{my_tree}{phylogenetic tree}

\item{ID}{Column name containing species names}

\item{Y}{Column name of dependent variable}

\item{X}{Column name of independent variable}

\item{method}{Eigenvector selection method, "ESRRV" or "mMorI", default is "mMorI"}

\item{estimator}{Estimator used in regression, "L1","L2","S","M" or "MM", default is "L2"}
}
\description{
Apply PVR using union eigenvector sets and robust esimators
}
