% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Select_EV_ESRRV.R
\name{Select_EV_ESRRV}
\alias{Select_EV_ESRRV}
\title{Select_EV_ESRRV}
\usage{
Select_EV_ESRRV(Y, pvr_df, my_tree)
}
\arguments{
\item{Y}{the traits used to select eigenvectors}

\item{pvr_df}{merged data frame,output of Merge_PVR_Traits}

\item{my_tree}{phylogenetic tree}
}
\description{
Select eigenvectors using ESRRV
}
