% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Union_EV.R
\name{Union_EV}
\alias{Union_EV}
\title{Union_EV}
\usage{
Union_EV(my_df, my_tree, ID, Y, X, method)
}
\arguments{
\item{my_df}{traits data frame}

\item{my_tree}{phylogenetic tree}

\item{ID}{Column name containing species names}

\item{Y}{Column name of dependent variable}

\item{X}{Column name of independent variable}

\item{method}{Eigenvector selection method, "ESRRV" or "mMorI"}
}
\description{
Get the union eigenvectors sets of X and Y
}
